Add some tests.
They just run the scanner on an NTFS and FAT32 disk image (1G each) and
compare the results to the expected output from previous runs.

This should at least let us know if something in the common paths
breaks or changes.

I need to make a proper set of disk images that exercises as many
codepaths as possible, but this is an MVP.